Move server code to `vs/server/node`
