+++ src/vs/server/node/remoteAgentEnvironmentImpl.ts
@@ -7,17 +7,17 @@ import { Event } from 'vs/base/common/event';
 import * as platform from 'vs/base/common/platform';
 import * as performance from 'vs/base/common/performance';
 import { URI } from 'vs/base/common/uri';
-import { createRemoteURITransformer } from 'vs/server/remoteUriTransformer';
+import { createRemoteURITransformer } from 'vs/server/node/remoteUriTransformer';
 import { IRemoteAgentEnvironmentDTO, IGetEnvironmentDataArguments, IScanExtensionsArguments, IScanSingleExtensionArguments } from 'vs/workbench/services/remote/common/remoteAgentEnvironmentChannel';
 import * as nls from 'vs/nls';
 import { FileAccess, Schemas } from 'vs/base/common/network';
-import { IServerEnvironmentService } from 'vs/server/serverEnvironmentService';
+import { IServerEnvironmentService } from 'vs/server/node/serverEnvironmentService';
 import { ExtensionScanner, ExtensionScannerInput, IExtensionResolver, IExtensionReference } from 'vs/workbench/services/extensions/node/extensionPoints';
 import { IServerChannel } from 'vs/base/parts/ipc/common/ipc';
 import { ExtensionIdentifier, IExtensionDescription } from 'vs/platform/extensions/common/extensions';
 import { transformOutgoingURIs } from 'vs/base/common/uriIpc';
 import { ILogService } from 'vs/platform/log/common/log';
-import { getNLSConfiguration, InternalNLSConfiguration } from 'vs/server/remoteLanguagePacks';
+import { getNLSConfiguration, InternalNLSConfiguration } from 'vs/server/node/remoteLanguagePacks';
 import { ContextKeyExpr, ContextKeyDefinedExpr, ContextKeyNotExpr, ContextKeyEqualsExpr, ContextKeyNotEqualsExpr, ContextKeyRegexExpr, IContextKeyExprMapper, ContextKeyExpression, ContextKeyInExpr, ContextKeyGreaterExpr, ContextKeyGreaterEqualsExpr, ContextKeySmallerExpr, ContextKeySmallerEqualsExpr } from 'vs/platform/contextkey/common/contextkey';
 import { listProcesses } from 'vs/base/node/ps';
 import { getMachineInfo, collectWorkspaceStats } from 'vs/platform/diagnostics/node/diagnosticsService';
