+++ src/vs/server/node/remoteExtensionHostAgentServer.ts
@@ -20,9 +20,9 @@ import * as platform from 'vs/base/common/platform';
 import { PersistentProtocol, ProtocolConstants } from 'vs/base/parts/ipc/common/ipc.net';
 import { NodeSocket, WebSocketNodeSocket } from 'vs/base/parts/ipc/node/ipc.net';
 import { ConnectionType, ConnectionTypeRequest, ErrorMessage, HandshakeMessage, IRemoteExtensionHostStartParams, ITunnelConnectionStartParams, SignRequest } from 'vs/platform/remote/common/remoteAgentConnection';
-import { ExtensionHostConnection } from 'vs/server/extensionHostConnection';
-import { ManagementConnection } from 'vs/server/remoteExtensionManagement';
-import { createRemoteURITransformer } from 'vs/server/remoteUriTransformer';
+import { ExtensionHostConnection } from 'vs/server/node/extensionHostConnection';
+import { ManagementConnection } from 'vs/server/node/remoteExtensionManagement';
+import { createRemoteURITransformer } from 'vs/server/node/remoteUriTransformer';
 import { ILogService, LogLevel, AbstractLogger, DEFAULT_LOG_LEVEL, MultiplexLogService, getLogLevel, LogService } from 'vs/platform/log/common/log';
 import { FileAccess, Schemas } from 'vs/base/common/network';
 import product from 'vs/platform/product/common/product';
@@ -54,8 +54,8 @@ import { IProductService } from 'vs/platform/product/common/productService';
 import { RemoteAgentConnectionContext } from 'vs/platform/remote/common/remoteAgentEnvironment';
 import { IPCServer, ClientConnectionEvent, IMessagePassingProtocol, StaticRouter, ProxyChannel } from 'vs/base/parts/ipc/common/ipc';
 import { Emitter, Event } from 'vs/base/common/event';
-import { RemoteAgentEnvironmentChannel } from 'vs/server/remoteAgentEnvironmentImpl';
-import { RemoteAgentFileSystemProviderChannel } from 'vs/server/remoteFileSystemProviderServer';
+import { RemoteAgentEnvironmentChannel } from 'vs/server/node/remoteAgentEnvironmentImpl';
+import { RemoteAgentFileSystemProviderChannel } from 'vs/server/node/remoteFileSystemProviderServer';
 import { REMOTE_FILE_SYSTEM_CHANNEL_NAME } from 'vs/workbench/services/remote/common/remoteFileSystemProviderClient';
 import { RequestChannel } from 'vs/platform/request/common/requestIpc';
 import { ExtensionManagementChannel } from 'vs/platform/extensionManagement/common/extensionManagementIpc';
@@ -63,28 +63,28 @@ import ErrorTelemetry from 'vs/platform/telemetry/node/errorTelemetry';
 import { ExtensionHostDebugBroadcastChannel } from 'vs/platform/debug/common/extensionHostDebugIpc';
 import { LogLevelChannel } from 'vs/platform/log/common/logIpc';
 import { IURITransformer } from 'vs/base/common/uriIpc';
-import { WebClientServer, serveError, serveFile } from 'vs/server/webClientServer';
+import { WebClientServer, serveError, serveFile } from 'vs/server/node/webClientServer';
 import { URI } from 'vs/base/common/uri';
 import { isEqualOrParent } from 'vs/base/common/extpath';
-import { IServerEnvironmentService, ServerEnvironmentService, ServerParsedArgs } from 'vs/server/serverEnvironmentService';
+import { IServerEnvironmentService, ServerEnvironmentService, ServerParsedArgs } from 'vs/server/node/serverEnvironmentService';
 import { basename, dirname, join } from 'vs/base/common/path';
 import { REMOTE_TERMINAL_CHANNEL_NAME } from 'vs/workbench/contrib/terminal/common/remoteTerminalChannel';
-import { RemoteTerminalChannel } from 'vs/server/remoteTerminalChannel';
+import { RemoteTerminalChannel } from 'vs/server/node/remoteTerminalChannel';
 import { LoaderStats } from 'vs/base/common/amd';
 import { RemoteExtensionLogFileName } from 'vs/workbench/services/remote/common/remoteAgentService';
 import { ExtensionManagementCLIService } from 'vs/platform/extensionManagement/common/extensionManagementCLIService';
 import { SpdLogLogger } from 'vs/platform/log/node/spdlogLog';
 import { IPtyService, TerminalSettingId } from 'vs/platform/terminal/common/terminal';
 import { PtyHostService } from 'vs/platform/terminal/node/ptyHostService';
-import { IRemoteTelemetryService, RemoteNullTelemetryService, RemoteTelemetryService } from 'vs/server/remoteTelemetryService';
+import { IRemoteTelemetryService, RemoteNullTelemetryService, RemoteTelemetryService } from 'vs/server/node/remoteTelemetryService';
 import { IUriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentity';
 import { UriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentityService';
 import { ICredentialsService } from 'vs/platform/credentials/common/credentials';
 import { CredentialsMainService } from 'vs/platform/credentials/node/credentialsMainService';
 import { IEncryptionService } from 'vs/workbench/services/encryption/common/encryptionService';
 import { EncryptionMainService } from 'vs/platform/encryption/node/encryptionMainService';
-import { RemoteTelemetryChannel } from 'vs/server/remoteTelemetryChannel';
-import { parseConnectionToken, ServerConnectionTokenParseError } from 'vs/server/connectionToken';
+import { RemoteTelemetryChannel } from 'vs/server/node/remoteTelemetryChannel';
+import { parseConnectionToken, ServerConnectionTokenParseError } from 'vs/server/node/connectionToken';
 
 const SHUTDOWN_TIMEOUT = 5 * 60 * 1000;
