+++ src/vs/server/node/remoteExtensionManagement.ts
