+++ src/vs/server/node/remoteFileSystemProviderServer.ts
@@ -8,11 +8,11 @@ import { URI, UriComponents } from 'vs/base/common/uri';
 import { IURITransformer } from 'vs/base/common/uriIpc';
 import { IFileChange } from 'vs/platform/files/common/files';
 import { ILogService } from 'vs/platform/log/common/log';
-import { createRemoteURITransformer } from 'vs/server/remoteUriTransformer';
+import { createRemoteURITransformer } from 'vs/server/node/remoteUriTransformer';
 import { RemoteAgentConnectionContext } from 'vs/platform/remote/common/remoteAgentEnvironment';
 import { DiskFileSystemProvider } from 'vs/platform/files/node/diskFileSystemProvider';
 import { posix, delimiter } from 'vs/base/common/path';
-import { IServerEnvironmentService } from 'vs/server/serverEnvironmentService';
+import { IServerEnvironmentService } from 'vs/server/node/serverEnvironmentService';
 import { AbstractDiskFileSystemProviderChannel, AbstractSessionFileWatcher, ISessionFileWatcher } from 'vs/platform/files/node/diskFileSystemProviderServer';
 import { IRecursiveWatcherOptions } from 'vs/platform/files/common/watcher';
