+++ src/vs/server/node/remoteTelemetryChannel.ts
@@ -8,7 +8,7 @@ import { Disposable } from 'vs/base/common/lifecycle';
 import { IServerChannel } from 'vs/base/parts/ipc/common/ipc';
 import { TelemetryLevel } from 'vs/platform/telemetry/common/telemetry';
 import { ITelemetryAppender } from 'vs/platform/telemetry/common/telemetryUtils';
-import { IRemoteTelemetryService } from 'vs/server/remoteTelemetryService';
+import { IRemoteTelemetryService } from 'vs/server/node/remoteTelemetryService';
 
 export class RemoteTelemetryChannel extends Disposable implements IServerChannel {
 	constructor(
