+++ src/vs/server/node/remoteUriTransformer.ts
@@ -6,7 +6,7 @@
 import { URITransformer, IURITransformer, IRawURITransformer } from 'vs/base/common/uriIpc';
 import { FileAccess } from 'vs/base/common/network';
 
-export const uriTransformerPath = FileAccess.asFileUri('vs/server/uriTransformer.js', require).fsPath;
+export const uriTransformerPath = FileAccess.asFileUri('vs/server/node/uriTransformer.js', require).fsPath;
 
 export function createRemoteURITransformer(remoteAuthority: string): IURITransformer {
 	const rawURITransformerFactory = <any>require.__$__nodeRequire(uriTransformerPath);
